Document `Valve`
