+++ src/faucets/mod.ts
@@ -1,3 +1,3 @@
 export * from './create-faucet.js';
 export * from './valve-closed.error.js';
-export * from './with-valve.js';
+export * from './valve.js';
